# docs/openapi/phase-2b.yaml
openapi: 3.1.0
info:
  title: ChangePath Phase 2b API Contract
  version: 0.1.0
  description: Section checkout and concurrency contract defined at the end of Phase 2a.
paths:
...
  /api/cases/{caseId}/assessment:
    put:
      summary: Save assessment answers with optimistic locking
      description: |
        Existing Phase 2a endpoint. In Phase 2b it will additionally validate
        section locks and reject stale writes with 409 Conflict responses.
      responses:
        '200':
          description: Save accepted
        '409':
          description: Stale assessment timestamp or section lock conflict
...